create base send details component
