+++ libs/tools/send/send-ui/src/send-form/components/send-form.component.html
@@ -1,10 +1,8 @@
 <form [id]="formId" [formGroup]="sendForm" [bitSubmit]="submit">
-  <!-- TODO: Should we show a loading spinner here? Or emit a ready event for the container to handle loading state -->
   <ng-container *ngIf="!loading">
-    <tools-send-text-details
-      *ngIf="config.sendType === SendType.Text"
+    <tools-send-details
       [config]="config"
       [originalSendView]="originalSendView"
-    ></tools-send-text-details>
+    ></tools-send-details>
   </ng-container>
 </form>
