+++ libs/tools/send/send-ui/src/send-form/send-form-container.ts
@@ -1,13 +1,13 @@
 import { SendView } from "@bitwarden/common/tools/send/models/view/send.view";
 
 import { SendFormConfig } from "./abstractions/send-form-config.service";
-import { SendTextDetailsComponent } from "./components/send-details/send-text-details.component";
+import { SendDetailsComponent } from "./components/send-details/send-details.component";
 /**
  * The complete form for a send. Includes all the sub-forms from their respective section components.
  * TODO: Add additional form sections as they are implemented.
  */
 export type SendForm = {
-  sendTextDetailsForm?: SendTextDetailsComponent["sendTextDetailsForm"];
+  sendDetailsForm?: SendDetailsComponent["sendDetailsForm"];
 };
 
 /**
